% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_about.R
\name{mod_about_ui}
\alias{mod_about_ui}
\alias{mod_about_server}
\title{mod_about_ui and mod_about_server}
\usage{
mod_about_ui(id)

mod_about_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
