% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_binder.R
\name{mod_binder_ui}
\alias{mod_binder_ui}
\alias{mod_binder_server}
\title{mod_binder_ui and mod_binder_server}
\usage{
mod_binder_ui(id)

mod_binder_server(input, output, session, img, r, ids)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
