% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dl.R
\name{mod_dl_ui}
\alias{mod_dl_ui}
\alias{mod_dl_server}
\title{mod_dl_ui and mod_dl_server}
\usage{
mod_dl_ui(id)

mod_dl_server(input, output, session, img)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
