% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_left.R
\name{mod_left_ui}
\alias{mod_left_ui}
\alias{mod_left_server}
\title{mod_left_ui and mod_left_server}
\usage{
mod_left_ui(id)

mod_left_server(input, output, session, img)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
