% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main.R
\name{mod_main_ui}
\alias{mod_main_ui}
\alias{mod_main_server}
\title{mod_main_ui and mod_main_server}
\usage{
mod_main_ui(id)

mod_main_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
