% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_my_hexes.R
\name{mod_my_hexes_ui}
\alias{mod_my_hexes_ui}
\alias{mod_my_hexes_server}
\title{mod_my_hexes_ui and mod_my_hexes_server}
\usage{
mod_my_hexes_ui(id)

mod_my_hexes_server(input, output, session, img)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
