% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pkg_name.R
\name{mod_pkg_name_ui}
\alias{mod_pkg_name_ui}
\title{mod_pkg_name_ui and mod_pkg_name_server}
\usage{
mod_pkg_name_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
