% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rendering.R
\name{mod_rendering_ui}
\alias{mod_rendering_ui}
\alias{mod_rendering_server}
\title{mod_rendering_ui and mod_rendering_server}
\usage{
mod_rendering_ui(id)

mod_rendering_server(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
