% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_restore.R
\name{mod_restore_ui}
\alias{mod_restore_ui}
\alias{mod_restore_server}
\title{mod_restore_ui and mod_restore_server}
\usage{
mod_restore_ui(id)

mod_restore_server(input, output, session, img, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
