% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_right.R
\name{mod_right_ui}
\alias{mod_right_ui}
\alias{mod_right_server}
\title{mod_right_ui and mod_right_server}
\usage{
mod_right_ui(id)

mod_right_server(input, output, session, img)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
