% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spotlight.R
\name{mod_spotlight_ui}
\alias{mod_spotlight_ui}
\alias{mod_spotlight_server}
\title{mod_spotlight_ui and mod_spotlight_server}
\usage{
mod_spotlight_ui(id)

mod_spotlight_server(input, output, session, img, r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
