% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(with_mongo = TRUE)
}
\arguments{
\item{with_mongo}{Do you want to use mongo as a backend?}
}
\description{
Run the Shiny Application
}
