% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hexa.R
\name{mod_hexa_ui}
\alias{mod_hexa_ui}
\title{mod_hexa_ui and mod_hexa_server}
\usage{
mod_hexa_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
